Common

Documentation

type TestNested = Reader [String] TestTree

A TestTree of tests under some name prefix.

runTestNestedIn :: [String] -> TestNested -> TestTree

Run a TestTree of tests with a given name prefix.

runTestNested :: TestNested -> TestTree

Run a TestTree of tests with an empty prefix.

testNested :: String -> [TestNested] -> TestNested

Descend into a name prefix.

goldenVsText :: TestName -> FilePath -> Text -> TestTree

Check the contents of a file against a Text.

goldenVsTextM :: TestName -> FilePath -> IO Text -> TestTree

Check the contents of a file against a Text.

goldenVsDoc :: TestName -> FilePath -> Doc ann -> TestTree

Check the contents of a file against a Doc.

goldenVsDocM :: TestName -> FilePath -> IO (Doc ann) -> TestTree

Check the contents of a file against a Doc.

nestedGoldenVsText :: TestName -> Text -> TestNested

Check the contents of a file under a name prefix against a Text.

nestedGoldenVsTextM :: TestName -> IO Text -> TestNested

Check the contents of a file under a name prefix against a Text.

nestedGoldenVsDoc :: TestName -> Doc ann -> TestNested

Check the contents of a file under a name prefix against a Text.

nestedGoldenVsDocM :: TestName -> IO (Doc ann) -> TestNested

Check the contents of a file under a name prefix against a Text.