Language.PlutusCore

Parser

parse :: ByteString -> Either (ParseError AlexPosn) (Program TyName Name AlexPosn)

Parse a ByteString containing a Plutus Core program, returning a ParseError if syntactically invalid.

>>> :set -XOverloadedStrings
>>> parse "(program 0.1.0 [(con addInteger) x y])"
Right (Program (AlexPn 1 1 2) (Version (AlexPn 9 1 10) 0 1 0) (Apply (AlexPn 15 1 16) (Apply (AlexPn 15 1 16) (Constant (AlexPn 17 1 18) (BuiltinName (AlexPn 21 1 22) AddInteger)) (Var (AlexPn 33 1 34) (Name {nameAttribute = AlexPn 33 1 34, nameString = "x", nameUnique = Unique {unUnique = 0}}))) (Var (AlexPn 35 1 36) (Name {nameAttribute = AlexPn 35 1 36, nameString = "y", nameUnique = Unique {unUnique = 1}}))))

parseST :: ByteString -> StateT IdentifierState (Except (ParseError AlexPosn)) (Program TyName Name AlexPosn)

parseTermST :: ByteString -> StateT IdentifierState (Except (ParseError AlexPosn)) (Term TyName Name AlexPosn)

parseTypeST :: ByteString -> StateT IdentifierState (Except (ParseError AlexPosn)) (Type TyName AlexPosn)

parseScoped :: (AsParseError e AlexPosn, AsValueRestrictionError e TyName AlexPosn, AsUniqueError e AlexPosn, MonadError e m, MonadQuote m) => ByteString -> m (Program TyName Name AlexPosn)

Parse and rewrite so that names are globally unique, not just unique within their scope.

parseProgram :: (AsParseError e AlexPosn, MonadError e m, MonadQuote m) => ByteString -> m (Program TyName Name AlexPosn)

Parse a PLC program. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseTerm :: (AsParseError e AlexPosn, MonadError e m, MonadQuote m) => ByteString -> m (Term TyName Name AlexPosn)

Parse a PLC term. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseType :: (AsParseError e AlexPosn, MonadError e m, MonadQuote m) => ByteString -> m (Type TyName AlexPosn)

Parse a PLC type. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

AST

data Term tyname name a

A Term is a value.

Constructors

Var a (name a)
A named variable

TyAbs a (tyname a) (Kind a) (Term tyname name a)

LamAbs a (name a) (Type tyname a) (Term tyname name a)

Apply a (Term tyname name a) (Term tyname name a)

Constant a (Constant a)
A constant term

Builtin a (Builtin a)

TyInst a (Term tyname name a) (Type tyname a)

Unwrap a (Term tyname name a)

IWrap a (Type tyname a) (Type tyname a) (Term tyname name a)

Error a (Type tyname a)

data Type tyname a

A Type assigned to expressions.

Constructors

TyVar a (tyname a)

TyFun a (Type tyname a) (Type tyname a)

TyIFix a (Type tyname a) (Type tyname a)
Fix-point type, for constructing self-recursive types

TyForall a (tyname a) (Kind a) (Type tyname a)

TyBuiltin a TypeBuiltin
Builtin type

TyInt a Natural
Type-level size

TyLam a (tyname a) (Kind a) (Type tyname a)

TyApp a (Type tyname a) (Type tyname a)

data Constant a

A constant value.

Constructors

BuiltinInt a Natural Integer

BuiltinBS a Natural ByteString

BuiltinSize a Natural

BuiltinStr a String

data Builtin a

Constructors

BuiltinName a BuiltinName

DynBuiltinName a DynamicBuiltinName

data Kind a

Kinds. Each type has an associated kind.

Constructors

Type a

KindArrow a (Kind a) (Kind a)

Size a

data ParseError a

An error encountered during parsing.

Constructors

LexErr String

Unexpected (Token a)

Overflow a Natural Integer

data Version a

Version of Plutus Core to be used for the program.

Constructors

Version a Natural Natural Natural

data Program tyname name a

A Program is simply a Term coupled with a Version of the core language.

Constructors

Program a (Version a) (Term tyname name a)

data Name a

A Name represents variables/names in Plutus Core.

Constructors

Name

Fields

nameAttribute :: a

nameString :: Text

The identifier name, for use in error messages.

nameUnique :: Unique

A Unique assigned to the name during lexing, allowing for cheap comparisons in the compiler.

newtype TyName a

We use a newtype to enforce separation between names used for types and those used for terms.

Constructors

TyName

Fields

unTyName :: Name a

newtype Unique

A unique identifier

Constructors

Unique

Fields

unUnique :: Int

type Size = Natural

type Value = Term

data BuiltinName

Builtin functions

Constructors

AddInteger

SubtractInteger

MultiplyInteger

DivideInteger

QuotientInteger

RemainderInteger

ModInteger

LessThanInteger

LessThanEqInteger

GreaterThanInteger

GreaterThanEqInteger

EqInteger

ResizeInteger

IntToByteString

Concatenate

TakeByteString

DropByteString

ResizeByteString

SHA2

SHA3

VerifySignature

EqByteString

TxHash

BlockNum

SizeOfInteger

newtype DynamicBuiltinName

The type of dynamic built-in functions. I.e. functions that exist on certain chains and do not exist on others. Each DynamicBuiltinName has an associated type and operational semantics -- this allows to type check and evaluate dynamic built-in names just like static ones.

Constructors

DynamicBuiltinName

Fields

unDynamicBuiltinName :: Text

The name of a dynamic built-in name.

data StagedBuiltinName

Either a BuiltinName (known statically) or a DynamicBuiltinName (known dynamically).

Constructors

StaticStagedBuiltinName BuiltinName

DynamicStagedBuiltinName DynamicBuiltinName

data TypeBuiltin

A builtin type

Constructors

TyByteString

TyInteger

TySize

TyString

newtype Normalized a

Constructors

Normalized

Fields

getNormalized :: a

type NormalizedType tyname a = Normalized (Type tyname a)

getNormalizedType :: NormalizedType tyname a -> Type tyname a

defaultVersion :: a -> Version a

The default version of Plutus Core supported by this library.

allBuiltinNames :: [BuiltinName]

The list of all BuiltinNames.

termLoc :: Term tyname name a -> a

tyLoc :: Type tyname a -> a

Lexer

data AlexPosn

Constructors

AlexPn !Int !Int !Int

Views

data IterApp head arg

A function (called "head") applied to a list of arguments (called "spine").

Constructors

IterApp

Fields

_iterAppHead :: head

_iterAppSpine :: [arg]

type TermIterApp tyname name a = IterApp (Term tyname name a) (Term tyname name a)

An iterated application of a Term to a list of Terms.

type PrimIterApp tyname name a = IterApp StagedBuiltinName (Value tyname name a)

An iterated application of a BuiltinName to a list of Values.

Formatting

format :: (AsParseError e AlexPosn, MonadError e m) => PrettyConfigPlc -> ByteString -> m Text

formatDoc :: (AsParseError e AlexPosn, MonadError e m) => PrettyConfigPlc -> ByteString -> m (Doc a)

Processing

newtype Gas

Constructors

Gas

Fields

unGas :: Natural

data ValueRestrictionError tyname a

Constructors

ValueRestrictionViolation a (tyname a)

class AsValueRestrictionError r tyname a | r -> tyname a where

Minimal complete definition

_ValueRestrictionError

Methods

_ValueRestrictionError :: Prism' r (ValueRestrictionError tyname a)

_ValueRestrictionViolation :: Prism' r (a, tyname a)

rename :: (Rename a, MonadQuote m) => a -> m a

Rename Uniques so that they're globally unique. In case there are any free variables, they must be left untouched. Must always assign new names to bound variables, so that rename can be used for alpha renaming as well.

Normalization

check :: Program tyname name a -> Maybe (NormalizationError tyname name a)

checkProgram :: (AsNormalizationError e TyName Name a, MonadError e m) => Program TyName Name a -> m ()

Ensure that all terms and types are well-formed accoring to Fig. 2

checkTerm :: (AsNormalizationError e TyName Name a, MonadError e m) => Term TyName Name a -> m ()

Ensure that all terms and types are well-formed accoring to Fig. 2

data NormalizationError tyname name a

checkFile :: FilePath -> IO (Maybe Text)

isTypeValue :: Type tyname a -> Bool

isTermValue :: Term tyname name a -> Bool

Check whether a Term is a Value. The term is assumed to be valid.

Type synthesis

typecheckProgram :: (AsTypeError e a, MonadError e m, MonadQuote m) => TypeConfig -> Program TyName Name a -> m (NormalizedType TyName ())

Type-check a program, returning a normalized type.

typecheckTerm :: (AsTypeError e a, MonadError e m, MonadQuote m) => TypeConfig -> Term TyName Name a -> m (NormalizedType TyName ())

Type-check a term, returning a normalized type.

kindCheck :: (AsTypeError e a, MonadError e m, MonadQuote m) => TypeConfig -> Type TyName a -> m (Kind ())

Kind-check a PLC type.

dynamicBuiltinNameMeaningsToTypes :: DynamicBuiltinNameMeanings -> DynamicBuiltinNameTypes

Extract the TypeScheme from a DynamicBuiltinNameMeaning and convert it to the corresponding Type TyName for each row of a DynamicBuiltinNameMeanings.

fileType :: FilePath -> IO Text

Given a file at fibonacci.plc, fileType "fibonacci.plc" will display its type or an error message.

fileNormalizeType :: Bool -> FilePath -> IO Text

fileTypeCfg :: PrettyConfigPlc -> FilePath -> IO Text

Given a file, display its type or an error message, optionally dumping annotations and debug information.

printType :: (AsParseError e AlexPosn, AsValueRestrictionError e TyName AlexPosn, AsUniqueError e AlexPosn, AsTypeError e AlexPosn, MonadError e m) => ByteString -> m Text

Print the type of a program contained in a ByteString

printNormalizeType :: (AsParseError e AlexPosn, AsValueRestrictionError e TyName AlexPosn, AsUniqueError e AlexPosn, AsTypeError e AlexPosn, MonadError e m) => Bool -> ByteString -> m Text

Print the type of a program contained in a ByteString

data InternalTypeError a

An internal error occurred during type checking.

Constructors

OpenTypeOfBuiltin (Type TyName ()) (Builtin ())

data TypeError a

Constructors

KindMismatch a (Type TyName ()) (Kind ()) (Kind ())

TypeMismatch a (Term TyName Name ()) (Type TyName ()) (NormalizedType TyName ())

UnknownDynamicBuiltinName a UnknownDynamicBuiltinNameError

InternalTypeErrorE a (InternalTypeError a)

FreeTypeVariableE (TyName a)

FreeVariableE (Name a)

OutOfGas

class AsTypeError r a | r -> a where

Minimal complete definition

_TypeError

Methods

_TypeError :: Prism' r (TypeError a)

_KindMismatch :: Prism' r (a, Type TyName (), Kind (), Kind ())

_TypeMismatch :: Prism' r (a, Term TyName Name (), Type TyName (), NormalizedType TyName ())

_UnknownDynamicBuiltinName :: Prism' r (a, UnknownDynamicBuiltinNameError)

_InternalTypeErrorE :: Prism' r (a, InternalTypeError a)

_FreeTypeVariableE :: Prism' r (TyName a)

_FreeVariableE :: Prism' r (Name a)

_OutOfGas :: Prism' r ()

data TypeConfig

Configuration of the type checker.

Constructors

TypeConfig

Fields

_typeConfigNormalize :: Bool

Whether to normalize type annotations.

_typeConfigDynBuiltinNameTypes :: DynamicBuiltinNameTypes

_typeConfigTyVarKinds :: TyVarKinds

_typeConfigVarTypes :: VarTypes

_typeConfigGas :: Maybe Gas

The upper limit on the length of type reductions. If set to Nothing, type reductions will be unbounded.

newtype DynamicBuiltinNameTypes

Mapping from DynamicBuiltinNames to their Types.

Constructors

DynamicBuiltinNameTypes

Fields

unDynamicBuiltinNameTypes :: Map DynamicBuiltinName (NormalizedType TyName ())

type TypeCheckM ann = ReaderT TypeConfig (ExceptT (TypeError ann) Quote)

The type checking monad contains the TypeConfig and it lets us throw TypeErrors.

parseTypecheck :: (AsParseError e AlexPosn, AsValueRestrictionError e TyName AlexPosn, AsUniqueError e AlexPosn, AsNormalizationError e TyName Name AlexPosn, AsTypeError e AlexPosn, MonadError e m, MonadQuote m) => TypeConfig -> ByteString -> m (NormalizedType TyName ())

Parse a program and typecheck it.

runTypeCheckM :: TypeConfig -> TypeCheckM a b -> ExceptT (TypeError a) Quote b

Run the type checker with a default context.

typecheckPipeline :: (AsNormalizationError e TyName Name a, AsTypeError e a, MonadError e m, MonadQuote m) => TypeConfig -> Program TyName Name a -> m (NormalizedType TyName ())

Typecheck a program.

defaultTypecheckerGas :: Maybe Gas

The default amount of gas to run the typechecker with.

defaultTypecheckerCfg :: TypeConfig

Errors

data Error a

Constructors

ParseErrorE (ParseError a)

ValueRestrictionErrorE (ValueRestrictionError TyName a)

UniqueCoherencyErrorE (UniqueError a)

TypeErrorE (TypeError a)

NormalizationErrorE (NormalizationError TyName Name a)

class AsError r a | r -> a where

Minimal complete definition

_Error

Methods

_Error :: Prism' r (Error a)

_ParseErrorE :: Prism' r (ParseError a)

_ValueRestrictionErrorE :: Prism' r (ValueRestrictionError TyName a)

_UniqueCoherencyErrorE :: Prism' r (UniqueError a)

_TypeErrorE :: Prism' r (TypeError a)

_NormalizationErrorE :: Prism' r (NormalizationError TyName Name a)

newtype UnknownDynamicBuiltinNameError

This error is returned whenever scope resolution of a DynamicBuiltinName fails.

Constructors

UnknownDynamicBuiltinNameErrorE DynamicBuiltinName

data UniqueError a

Constructors

MultiplyDefined Unique a a

IncoherentUsage Unique a a

FreeVariable Unique a

Base functors

data TermF tyname name a x

Constructors

VarF a (name a)

TyAbsF a (tyname a) (Kind a) x

LamAbsF a (name a) (Type tyname a) x

ApplyF a x x

ConstantF a (Constant a)

BuiltinF a (Builtin a)

TyInstF a x (Type tyname a)

UnwrapF a x

IWrapF a (Type tyname a) (Type tyname a) x

ErrorF a (Type tyname a)

data TypeF tyname a x

Constructors

TyVarF a (tyname a)

TyFunF a x x

TyIFixF a x x

TyForallF a (tyname a) (Kind a) x

TyBuiltinF a TypeBuiltin

TyIntF a Natural

TyLamF a (tyname a) (Kind a) x

TyAppF a x x

Quotation and term construction

type Quote = QuoteT Identity

A non-transformer version of QuoteT.

runQuote :: Quote a -> a

See runQuoteT.

data QuoteT m a

The "quotation" monad transformer. Within this monad you can do safe construction of PLC terms using quasiquotation, fresh-name generation, and parsing.

runQuoteT :: Monad m => QuoteT m a -> m a

Run a quote from an empty identifier state. Note that the resulting term cannot necessarily be safely combined with other terms - that should happen inside QuoteT.

class Monad m => MonadQuote m

A monad that allows lifting of quoted expressions.

liftQuote :: MonadQuote m => Quote a -> m a

Name generation

freshUnique :: MonadQuote m => m Unique

Get a fresh Unique.

freshName :: Monad m => a -> Text -> QuoteT m (Name a)

Get a fresh Name, given the annotation and the Text name.

freshTyName :: Monad m => a -> Text -> QuoteT m (TyName a)

Get a fresh TyName, given the annotation and the Text name.

Quasi-Quoters

plcType :: QuasiQuoter

A quasiquoter for creating Plutus Core types.

plcTerm :: QuasiQuoter

A quasiquoter for creating Plutus Core terms.

plcProgram :: QuasiQuoter

A quasiquoter for creating Plutus Core programs.

Evaluation

data EvaluationResultF a

The parameterized type of results various evaluation engines return.

Constructors

EvaluationSuccess a

EvaluationFailure

type EvaluationResult = EvaluationResultF (Value TyName Name ())

The type of results various evaluation engines return.

Combining programs

applyProgram :: Program tyname name () -> Program tyname name () -> Program tyname name ()

Take one PLC program and apply it to another.