Language.PlutusCore.Constant

Description

Reexports from modules from the Constant folder.

Documentation

data BuiltinSized

Built-in types indexed by size.

Constructors

BuiltinSizedInt

BuiltinSizedBS

BuiltinSizedSize

data TypedBuiltinSized a where

Built-in types indexed by size along with their denotation.

Constructors

TypedBuiltinSizedInt :: TypedBuiltinSized Integer

TypedBuiltinSizedBS :: TypedBuiltinSized ByteString

TypedBuiltinSizedSize :: TypedBuiltinSized ()

data SizeEntry size

Type-level sizes.

Constructors

SizeValue Size
A constant size.

SizeBound size
A bound size variable.

data BuiltinType size

Built-in types.

Constructors

BuiltinSized (SizeEntry size) BuiltinSized

BuiltinBool

data TypedBuiltin size a where

Built-in types. A type is considired "built-in" if it can appear in the type signature of a primitive operation. So boolean is considered built-in even though it is defined in PLC and is not primitive.

Constructors

TypedBuiltinSized :: SizeEntry size -> TypedBuiltinSized a -> TypedBuiltin size a

TypedBuiltinBool :: TypedBuiltin size Bool

TypedBuiltinDyn :: KnownDynamicBuiltinType dyn => TypedBuiltin size dyn

data TypedBuiltinValue size a

A TypedBuiltin packaged together with a value of the type that the TypedBuiltin denotes.

Constructors

TypedBuiltinValue (TypedBuiltin size a) a

data TypeScheme size a r where

Type schemes of primitive operations. a is the Haskell denotation of a PLC type represented as a TypeScheme. r is the resulting type in a, e.g. the resulting type in ByteString -> Size -> Integer is Integer.

Constructors

TypeSchemeBuiltin :: TypedBuiltin size a -> TypeScheme size a a

TypeSchemeArrow :: TypeScheme size a q -> TypeScheme size b r -> TypeScheme size (a -> b) r infixr 9

TypeSchemeAllSize :: (size -> TypeScheme size a r) -> TypeScheme size a r

data TypedBuiltinName a r

A BuiltinName with an associated TypeScheme.

Constructors

TypedBuiltinName BuiltinName (forall size. TypeScheme size a r)

data DynamicBuiltinNameMeaning

The meaning of a dynamic built-in name consists of its Type represented as a TypeScheme and its Haskell denotation.

Constructors

DynamicBuiltinNameMeaning (forall size. TypeScheme size a r) a

data DynamicBuiltinNameDefinition

The definition of a dynamic built-in consists of its name and meaning.

Constructors

DynamicBuiltinNameDefinition DynamicBuiltinName DynamicBuiltinNameMeaning

newtype DynamicBuiltinNameMeanings

Mapping from DynamicBuiltinNames to their DynamicBuiltinNameMeanings.

Constructors

DynamicBuiltinNameMeanings

Fields

unDynamicBuiltinNameMeanings :: Map DynamicBuiltinName DynamicBuiltinNameMeaning

type Evaluator f m = DynamicBuiltinNameMeanings -> f TyName Name () -> m EvaluationResult

type Evaluate m = ReaderT (Evaluator Term m) m

class KnownDynamicBuiltinType dyn where

Haskell types known to exist on the PLC side.

Minimal complete definition

getTypeEncoding, makeDynamicBuiltin, readDynamicBuiltin

Methods

getTypeEncoding :: proxy dyn -> Quote (Type TyName ())

The type representing dyn used on the PLC side.

makeDynamicBuiltin :: dyn -> Quote (Maybe (Term TyName Name ()))

Convert a Haskell value to the corresponding PLC value. Nothing represents a conversion failure.

readDynamicBuiltin :: Monad m => Evaluator Term m -> Term TyName Name () -> m (Maybe dyn)

Convert a PLC value to the corresponding Haskell value. Nothing represents a conversion failure.

eraseTypedBuiltinSized :: TypedBuiltinSized a -> BuiltinSized

Convert a TypedBuiltinSized to its untyped counterpart.

runEvaluate :: Evaluator Term m -> Evaluate m a -> m a

withEvaluator :: (Evaluator Term m -> m a) -> Evaluate m a

readDynamicBuiltinM :: (Monad m, KnownDynamicBuiltinType dyn) => Term TyName Name () -> Evaluate m (Maybe dyn)

typedAddInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer

Typed AddInteger.

typedSubtractInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer

Typed SubtractInteger.

typedMultiplyInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer

Typed MultiplyInteger.

typedDivideInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer

Typed DivideInteger.

typedQuotientInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer

Typed QuotientInteger

typedModInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer

Typed ModInteger

typedRemainderInteger :: TypedBuiltinName (Integer -> Integer -> Integer) Integer

Typed RemainderInteger.

typedLessThanInteger :: TypedBuiltinName (Integer -> Integer -> Bool) Bool

Typed LessThanInteger.

typedLessThanEqInteger :: TypedBuiltinName (Integer -> Integer -> Bool) Bool

Typed LessThanEqInteger.

typedGreaterThanInteger :: TypedBuiltinName (Integer -> Integer -> Bool) Bool

Typed GreaterThanInteger.

typedGreaterThanEqInteger :: TypedBuiltinName (Integer -> Integer -> Bool) Bool

Typed GreaterThanEqInteger.

typedEqInteger :: TypedBuiltinName (Integer -> Integer -> Bool) Bool

Typed EqInteger.

typedResizeInteger :: TypedBuiltinName (() -> Integer -> Integer) Integer

Typed ResizeInteger.

typedIntToByteString :: TypedBuiltinName (() -> Integer -> ByteString) ByteString

Typed IntToByteString.

typedConcatenate :: TypedBuiltinName (ByteString -> ByteString -> ByteString) ByteString

Typed Concatenate.

typedTakeByteString :: TypedBuiltinName (Integer -> ByteString -> ByteString) ByteString

Typed TakeByteString.

typedDropByteString :: TypedBuiltinName (Integer -> ByteString -> ByteString) ByteString

Typed DropByteString.

typedSHA2 :: TypedBuiltinName (ByteString -> ByteString) ByteString

Typed SHA2.

typedSHA3 :: TypedBuiltinName (ByteString -> ByteString) ByteString

Typed SHA3.

typedVerifySignature :: TypedBuiltinName (ByteString -> ByteString -> ByteString -> Bool) Bool

Typed VerifySignature.

typedResizeByteString :: TypedBuiltinName (() -> ByteString -> ByteString) ByteString

Typed ResizeByteString.

typedEqByteString :: TypedBuiltinName (ByteString -> ByteString -> Bool) Bool

Typed EqByteString.

typedTxHash :: TypedBuiltinName ByteString ByteString

Typed TxHash.

typedBlockNum :: TypedBuiltinName (() -> Integer) Integer

Typed BlockNum.

typedSizeOfInteger :: TypedBuiltinName (Integer -> ()) ()

Typed SizeOfInteger.

builtinNameAsTerm :: BuiltinName -> Term tyname name ()

Lift a BuiltinName to Term.

dynamicBuiltinNameAsTerm :: DynamicBuiltinName -> Term tyname name ()

Lift a DynamicBuiltinName to Term.

toBoundsInt :: Size -> (Integer, Integer)

Return the [-2^(8s - 1), 2^(8s - 1)) bounds for integers of a given Size.

toInclusiveBoundsInt :: Size -> (Integer, Integer)

Return the [-2^(8s - 1), 2^(8s - 1) - 1] bounds for integers of a given Size.

checkBoundsInt :: Size -> Integer -> Bool

Check whether an Integer is in the [-2^(8s - 1), 2^(8s - 1)) interval.

checkBoundsBS :: Size -> ByteString -> Bool

Check whether the length of a ByteString is less than or equal to a given Size.

sizeOfInteger :: Integer -> Size

Compute the size of an Integer. See also toBoundsInt.

sizeOfByteString :: ByteString -> Size

Compute the size of a ByteString. See also toBoundsBS.

makeAutoSizedBuiltinInt :: Integer -> Constant ()

Make a Constant out of an Integer. The size is computed using sizeOfInteger.

makeAutoSizedBuiltinBS :: ByteString -> Constant ()

Make a Constant out of a ByteString. The size is computed using sizeOfBS.

makeDynBuiltinInt

Arguments

:: Type tyname ()
An actual size s.

-> Term tyname name ()
A singleton for s.

-> Integer
An Integer to lift.

-> Term tyname name ()

Convert a Haskell Integer to the corresponding PLC integer.

makeDynBuiltinIntSizedAs

Arguments

:: Type tyname ()
An actual size s.

-> Term tyname name ()
An integer to take the singleton size from.

-> Integer
An Integer to lift.

-> Term tyname name ()

Convert a Haskell Integer to the corresponding PLC integer, taking the size singleton from an already existing PLC integer.

makeBuiltinInt :: Size -> Integer -> Maybe (Constant ())

Check whether an Integer is in bounds (see checkBoundsInt) and return it as a Constant.

makeBuiltinBS :: Size -> ByteString -> Maybe (Constant ())

Check whether a ByteString is in bounds (see checkBoundsBS) and return it as a Constant.

makeBuiltinStr :: String -> Constant ()

makeSizedConstant :: Size -> TypedBuiltinSized a -> a -> Maybe (Constant ())

Convert a Haskell value to the corresponding PLC constant indexed by size checking all constraints (e.g. an Integer is in appropriate bounds) along the way.

makeBuiltinBool :: Bool -> Quote (Term TyName Name ())

Convert a Bool to the corresponding PLC's boolean.

makeBuiltin :: TypedBuiltinValue Size a -> Quote (Maybe (Term TyName Name ()))

Convert a Haskell value to the corresponding PLC value checking all constraints (e.g. an Integer is in appropriate bounds) along the way.

unsafeMakeBuiltin :: PrettyDynamic a => TypedBuiltinValue Size a -> Quote (Term TyName Name ())

Convert a Haskell value to a PLC value checking all constraints (e.g. an Integer is in appropriate bounds) along the way and fail in case constraints are not satisfied.

unsafeMakeDynamicBuiltin :: (KnownDynamicBuiltinType dyn, PrettyDynamic dyn) => dyn -> Quote (Term TyName Name ())

Convert a Haskell value to a PLC value of a dynamic built-in type.

makeSizedConstantNOCHECK :: Size -> TypedBuiltinSized a -> a -> Constant ()

Convert a Haskell value to the corresponding PLC constant indexed by size without checking constraints (e.g. an Integer is in appropriate bounds). This function allows to fake a Constant with a wrong size and thus it's highly unsafe and should be used with great caution.

makeBuiltinNOCHECK :: PrettyDynamic a => TypedBuiltinValue Size a -> Quote (Term TyName Name ())

Convert a Haskell value to the corresponding PLC value without checking constraints (e.g. an Integer is in appropriate bounds). This function allows to fake a Term with a wrong size and thus it's highly unsafe and should be used with great caution.

flattenSizeEntry :: SizeEntry Size -> Size

Extract the Size from a SizeEntry.

eraseTypedBuiltinSized :: TypedBuiltinSized a -> BuiltinSized

Convert a TypedBuiltinSized to its untyped counterpart.

mapSizeEntryTypedBuiltin :: (SizeEntry size -> SizeEntry size') -> TypedBuiltin size a -> TypedBuiltin size' a

Alter the SizeEntry of a TypedBuiltin.

mapSizeTypedBuiltin :: (size -> size') -> TypedBuiltin size a -> TypedBuiltin size' a

Alter the size of a TypedBuiltin size.

closeTypedBuiltin :: TypedBuiltin Size a -> TypedBuiltin b a

Map each SizeBound to SizeValue.

typedBuiltinSizedToType :: TypedBuiltinSized a -> Type TyName ()

Convert a TypedBuiltinSized to the corresponding TypeBuiltin and wrap the result in TyBuiltin to get a Type.

withTypedBuiltinSized :: BuiltinSized -> (forall a. TypedBuiltinSized a -> c) -> c

Apply a continuation to the typed version of a BuiltinSized.

withTypedBuiltin :: BuiltinType size -> (forall a. TypedBuiltin size a -> c) -> c

Apply a continuation to the typed version of a Builtin.

typeSchemeResult :: TypeScheme () a r -> TypedBuiltin () r

The resulting TypedBuiltin of a TypeScheme.

typedBuiltinToType :: TypedBuiltin (Type TyName ()) a -> Quote (Type TyName ())

Convert a TypedBuiltin to the corresponding Type.

typeSchemeToType :: TypeScheme (Type TyName ()) a r -> Quote (Type TyName ())

Convert a TypeScheme to the corresponding Type. Basically, a map from the PHOAS representation to the FOAS one.

dynamicBuiltinNameMeaningToType :: DynamicBuiltinNameMeaning -> Quote (Type TyName ())

Extract the TypeScheme from a DynamicBuiltinNameMeaning and convert it to the corresponding Type.

insertDynamicBuiltinNameDefinition :: DynamicBuiltinNameDefinition -> DynamicBuiltinNameMeanings -> DynamicBuiltinNameMeanings

Insert a DynamicBuiltinNameDefinition into a DynamicBuiltinNameMeanings.

withTypedBuiltinName :: BuiltinName -> (forall a r. TypedBuiltinName a r -> c) -> c

Apply a continuation to the typed version of a BuiltinName.

typeOfTypedBuiltinName :: TypedBuiltinName a r -> Quote (Type TyName ())

Return the Type of a TypedBuiltinName.

typeOfBuiltinName :: BuiltinName -> Quote (Type TyName ())

Return the Type of a BuiltinName.

class PrettyDynamic a where

Same as the Pretty class, but is specifically for dynamic built-in types as their pretty-printing can be rather weird (see the PrettyDynamic BSL.ByteString instance for example).

Methods

prettyDynamic :: a -> Doc ann

prettyDynamic :: Pretty a => a -> Doc ann

data ConstAppError

The type of constant applications errors.

Constructors

SizeMismatchConstAppError Size (Constant ())
A mismatch between expected and actual sizes.

IllTypedConstAppError BuiltinSized (Constant ())
A mismatch between the type of an argument function expects and its actual type.

ExcessArgumentsConstAppError [Value TyName Name ()]
A constant is applied to more arguments than needed in order to reduce. Note that this error occurs even if an expression is well-typed, because constant application is supposed to be computed as soon as there are enough arguments.

SizedNonConstantConstAppError (Value TyName Name ())
An argument of a sized type is not a constant.

UnreadableBuiltinConstAppError (Value TyName Name ())
Could not construct denotation for a built-in.

data ConstAppResult

The type of constant applications results.

Constructors

ConstAppSuccess (Value TyName Name ())
Successfully computed a value.

ConstAppFailure
Not enough gas.

ConstAppStuck
Not enough arguments.

ConstAppError ConstAppError
An internal error occurred during evaluation.

makeConstAppResult :: TypedBuiltinValue Size a -> Quote ConstAppResult

Same as makeBuiltin, but returns a ConstAppResult.

applyTypeSchemed :: Monad m => TypeScheme SizeVar a r -> a -> [Value TyName Name ()] -> QuoteT (Evaluate m) ConstAppResult

Apply a function with a known TypeScheme to a list of Constants (unwrapped from Values). Checks that the constants are of expected types and there are no size mismatches.

applyBuiltinName :: Monad m => BuiltinName -> [Value TyName Name ()] -> QuoteT (Evaluate m) ConstAppResult

Apply a TypedBuiltinName to a list of Constants (unwrapped from Values) Checks that the constants are of expected types and there are no size mismatches.