Language.PlutusCore.Constant.Dynamic

Description

Reexports from modules from the Dynamic folder.

Documentation

newtype OnChain (names :: [Symbol]) f (tyname :: * -> *) (name :: * -> *) ann

Constructors

OnChain

Fields

unOnChain :: f tyname name ann

type OnChainEvaluator names f r = DynamicBuiltinNameMeanings -> OnChain names f TyName Name () -> r

The type of evaluators that run on-chain. This is very similar to Evaluator, except an OnChainEvaluator is allowed to return any r, not just a m EvaluationResult for some m. This is because an evaluator running on-chain can perform arbitrary effects and return an arbitrary result containing an EvaluationResult somewhere deep inside r.

type OnChainTransformer f names r names' r' = OnChainEvaluator names f r -> OnChainEvaluator names' f r'

The type of functions that transform OnChainEvaluators.

type OnChainHandler name f r s = forall names. OnChainTransformer f names r (name ': names) s

The type of handlers of outermost dynamic built-in names.

mangleOnChain :: OnChain names f tyname name ann -> OnChain names' f tyname name ann

handleDynamicByMeaning :: forall name f r. KnownSymbol name => DynamicBuiltinNameMeaning -> OnChainHandler name f r r

Interpret a DynamicBuiltinNameMeaning as an OnChainHandler that doesn't change the resulting type of evaluation.

dynamicEmit :: Term tyname name ()

handleDynamicEmit :: OnChainHandler "emit" f r (forall a. (forall size. TypedBuiltin size a) -> IO ([a], r))

dynamicLog :: Term tyname name ()

handleDynamicLog :: OnChainHandler "log" f r (IO ([String], r))

evaluateHandlersBy :: Evaluator f m -> (Evaluator (OnChain '[] f) m -> OnChainEvaluator names f r) -> OnChain names f TyName Name () -> r

newtype PlcList a

Constructors

PlcList

Fields

unPlcList :: [a]

withEmit :: ((a -> IO ()) -> IO b) -> IO ([a], b)

newtype EmitHandler r

Constructors

EmitHandler

Fields

unEmitHandler :: DynamicBuiltinNameMeanings -> Term TyName Name () -> IO r

feedEmitHandler :: Term TyName Name () -> EmitHandler r -> IO r

withEmitHandler :: Evaluator Term m -> (EmitHandler (m EvaluationResult) -> IO r2) -> IO r2

withEmitTerm :: (forall size. TypedBuiltin size a) -> (Term TyName Name () -> EmitHandler r1 -> IO r2) -> EmitHandler r1 -> IO ([a], r2)

withEmitEvaluateBy :: Evaluator Term m -> (forall size. TypedBuiltin size a) -> (Term TyName Name () -> Term TyName Name ()) -> IO ([a], m EvaluationResult)

dynamicCallAssign :: (forall size. TypedBuiltin size a) -> DynamicBuiltinName -> (a -> IO ()) -> DynamicBuiltinNameDefinition

dynamicCallTypeScheme :: (forall size. TypedBuiltin size a) -> forall size. TypeScheme size (a -> ()) ()

dynamicCall :: DynamicBuiltinName -> Term tyname name ()

dynamicCharToStringName :: DynamicBuiltinName

dynamicCharToStringMeaning :: DynamicBuiltinNameMeaning

dynamicCharToStringDefinition :: DynamicBuiltinNameDefinition

dynamicCharToString :: Term tyname name ()

dynamicAppendName :: DynamicBuiltinName

dynamicAppendMeaning :: DynamicBuiltinNameMeaning

dynamicAppendDefinition :: DynamicBuiltinNameDefinition

dynamicAppend :: Term tyname name ()

dynamicTraceName :: DynamicBuiltinName