Language.PlutusCore.DeBruijn

Description

Support for using de Bruijn indices for term and type names.

Documentation

newtype Index

A relative index used for de Bruijn identifiers.

Constructors

Index Natural

data DeBruijn a

A term name as a de Bruijn index.

Constructors

DeBruijn

Fields

dbnAttribute :: a

dbnString :: Text

dbnIndex :: Index

newtype TyDeBruijn a

A type name as a de Bruijn index.

Constructors

TyDeBruijn (DeBruijn a)

data FreeVariableError

We cannot do a correct translation to or from de Bruijn indices if the program is not well-scoped. So we throw an error in such a case.

Constructors

FreeUnique Unique

FreeIndex Index

deBruijnTy :: MonadError FreeVariableError m => Type TyName a -> m (Type TyDeBruijn a)

Convert a Type with TyNames into a Type with TyDeBruijns.

deBruijnTerm :: MonadError FreeVariableError m => Term TyName Name a -> m (Term TyDeBruijn DeBruijn a)

Convert a Term with TyNames and Names into a Term with TyDeBruijns and DeBruijns.

deBruijnProgram :: MonadError FreeVariableError m => Program TyName Name a -> m (Program TyDeBruijn DeBruijn a)

Convert a Program with TyNames and Names into a Program with TyDeBruijns and DeBruijns.

unDeBruijnTy :: (MonadQuote m, MonadError FreeVariableError m) => Type TyDeBruijn a -> m (Type TyName a)

Convert a Type with TyDeBruijns into a Type with TyNames.

unDeBruijnTerm :: (MonadQuote m, MonadError FreeVariableError m) => Term TyDeBruijn DeBruijn a -> m (Term TyName Name a)

Convert a Term with TyDeBruijns and DeBruijns into a Term with TyNames and Names.

unDeBruijnProgram :: (MonadQuote m, MonadError FreeVariableError m) => Program TyDeBruijn DeBruijn a -> m (Program TyName Name a)

Convert a Program with TyDeBruijns and DeBruijns into a Program with TyNames and Names.