Language.PlutusCore.Evaluation.CkMachine

Description

The CK machine.

Documentation

type CkMachineException = MachineException NoDynamicBuiltinNamesMachineError

The CK machine-specific MachineException.

data EvaluationResultF a

The parameterized type of results various evaluation engines return.

Constructors

EvaluationSuccess a

EvaluationFailure

type EvaluationResult = EvaluationResultF (Value TyName Name ())

The type of results various evaluation engines return.

applyEvaluateCkBuiltinName :: BuiltinName -> [Value TyName Name ()] -> ConstAppResult

evaluateCk :: Term TyName Name () -> EvaluationResult

Evaluate a term using the CK machine. May throw a CkMachineException. This differs from the spec version: we do not have the following rule:

s , {_ A} ◁ F ↦ s ◁ W  -- Fully saturated constant, {F A} ~> W.

The reason for that is that the operational semantics of constant applications is unaffected by types as it supports full type erasure, hence {F A} can never compute if F does not compute, so we simply do not introduce a rule that can't possibly fire.

runCk :: Program TyName Name () -> EvaluationResult

Run a program using the CK machine. May throw a CkMachineException. Calls evaluateCk under the hood, so the same caveats apply.