Language.PlutusCore.Evaluation.MachineException

Description

The exceptions that an abstract machine can throw.

Documentation

data MachineError err

Errors which can occur during a run of an abstract machine.

Constructors

NonPrimitiveInstantiationMachineError
An attempt to reduce a not immediately reducible type instantiation.

NonWrapUnwrappedMachineError
An attempt to unwrap a not wrapped term.

NonPrimitiveApplicationMachineError
An attempt to reduce a not immediately reducible application.

OpenTermEvaluatedMachineError
An attempt to evaluate an open term.

ConstAppMachineError ConstAppError
An attempt to compute a constant application resulted in ConstAppError.

OtherMachineError err

data MachineException err

The type of exceptions an abstract machine can throw.

Constructors

MachineException

Fields

_machineExceptionError :: MachineError err

An error.

_machineExceptionCause :: Term TyName Name ()

A Term that caused the error.