Language.PlutusCore.Generators

Description

Reexports from modules from the Language.PlutusCore.Generators.Internal folder.

Documentation

liftT :: (MFunctor t, MonadTrans s, Monad m) => t m a -> t (s m) a

hoist lift

hoistSupply :: (MFunctor t, Monad m) => r -> t (ReaderT r m) a -> t m a

Supply an environment to an inner ReaderT.

choiceDef :: Monad m => GenT m a -> [GenT m a] -> GenT m a

Same as choice, but with a default generator to be used when the supplied list of generators is empty.

forAllNoShow :: Monad m => Gen a -> PropertyT m a

Generate a value, but do not show it in case an error occurs.

forAllNoShowT :: Monad m => GenT m a -> PropertyT m a

Generate a value, but do not show it in case an error occurs. A supplied generator has access to the Monad the whole property has access to.

forAllPretty :: (Monad m, Pretty a) => Gen a -> PropertyT m a

Generate a value using the Pretty class for getting its String representation.

forAllPrettyT :: (Monad m, Pretty a) => GenT m a -> PropertyT m a

Generate a value using the Pretty class for getting its String representation. A supplied generator has access to the Monad the whole property has access to.

forAllPrettyPlc :: (Monad m, PrettyPlc a) => Gen a -> PropertyT m a

Generate a value using the PrettyPlc constraint for getting its String representation.

forAllPrettyPlcT :: (Monad m, PrettyPlc a) => GenT m a -> PropertyT m a

Generate a value using the PrettyPlc constraint for getting its String representation. A supplied generator has access to the Monad the whole property has access to.

forAllPrettyPlcMaybe :: (Monad m, PrettyPlc a) => Gen (Maybe a) -> PropertyT m (Maybe a)

Generate a value wrapped in Maybe using the PrettyPlc constraint for getting its String representation.

forAllPrettyPlcMaybeT :: (Monad m, PrettyPlc a) => GenT m (Maybe a) -> PropertyT m (Maybe a)

Generate a value wrapped in Maybe using the PrettyPlc constraint for getting its String representation. A supplied generator has access to the Monad the whole property has access to.

runQuoteSampleSucceed :: GenT Quote (Maybe a) -> IO a

Run a generator until it succeeds with a Just.

errorPlc :: PrettyPlc err => err -> b

Throw a PLC error.

data TypeEvalCheckError

The type of errors that can occur during type-eval checking.

Constructors

TypeEvalCheckErrorIllFormed (Error ())

TypeEvalCheckErrorIllEvaled (Value TyName Name ()) (Value TyName Name ())
The former is an expected result of evaluation, the latter -- is an actual one.

data TypeEvalCheckResult

Type-eval checking of a term results in a value of this type.

Constructors

TypeEvalCheckResult

Fields

_termCheckResultType :: NormalizedType TyName ()

The type of the term.

_termCheckResultValue :: EvaluationResult

The result of evaluation of the term.

type TypeEvalCheckM = ExceptT TypeEvalCheckError Quote

The monad type-eval checking runs in.

typeEvalCheckBy

Arguments

:: (Term TyName Name () -> EvaluationResult)
An evaluator.

-> TermOf (TypedBuiltinValue Size a)

-> TypeEvalCheckM (TermOf TypeEvalCheckResult)

Type check and evaluate a term and check that the expected result is equal to the actual one.

unsafeTypeEvalCheck :: forall a. TermOf (TypedBuiltinValue Size a) -> Quote (Maybe (TermOf (Value TyName Name ())))

Type check and evaluate a term and check that the expected result is equal to the actual one. Throw an error in case something goes wrong.

data TermOf a

A PLC Term along with the correspoding Haskell value.

Constructors

TermOf

Fields

_termOfTerm :: Term TyName Name ()

The PLC term

_termOfValue :: a

The Haskell value.

type TypedBuiltinGenT m = forall a. TypedBuiltin Size a -> GenT m (TermOf a)

A function of this type generates values of built-in typed (see TypedBuiltin for the list of such types) and returns it along with the corresponding PLC value. Bounds induced (as per the spec) by the Size values must be met, but can be narrowed.

type TypedBuiltinGen = TypedBuiltinGenT Identity

TypedBuiltinGenT specified to Identity.

genLowerBytes :: Monad m => Range Int -> GenT m ByteString

Generate a UTF-8 lazy ByteString containg lower-case letters.

updateTypedBuiltinGenInt :: MonadQuote m => (Integer -> Integer -> GenT m Integer) -> TypedBuiltinGenT m -> TypedBuiltinGenT m

Update a typed built-ins generator by overwriting the integers generator.

updateTypedBuiltinGenBS :: MonadQuote m => (Int -> GenT m ByteString) -> TypedBuiltinGenT m -> TypedBuiltinGenT m

Update a typed built-ins generator by overwriting the bytestrings generator.

updateTypedBuiltinGenSize :: MonadQuote m => TypedBuiltinGenT m -> TypedBuiltinGenT m

Update a typed built-ins generator by overwriting the sizes generator.

updateTypedBuiltinGenBool :: MonadQuote m => GenT m Bool -> TypedBuiltinGenT m -> TypedBuiltinGenT m

Update a typed built-ins generator by overwriting the booleans generator.

genTypedBuiltinFail :: Monad m => TypedBuiltinGenT m

A built-ins generator that always fails.

genTypedBuiltinDef :: MonadQuote m => TypedBuiltinGenT m

A default sized built-ins generator that produces values in bounds seen in the spec.

genTypedBuiltinOutOfBounds :: MonadQuote m => TypedBuiltinGenT m

A sized built-ins generator that produces values outside of bounds seen in the spec for integers and bytestrings.

genTypedBuiltinSmall :: MonadQuote m => TypedBuiltinGenT m

A sized built-ins generator that produces small values in bounds seen in the spec.

genTypedBuiltinSum :: MonadQuote m => TypedBuiltinGenT m

A sized built-ins generator that produces Integers in bounds narrowed by a factor of 2, so that one can use '(+)' or '(-)' over such integers without the risk of getting an overflow.

genTypedBuiltinMultiply :: MonadQuote m => TypedBuiltinGenT m

A sized built-ins generator that produces Integers in bounds narrowed by isqrtt, so that one can use '(*)' over such integers without the risk of getting an overflow.

genTypedBuiltinDivide :: MonadQuote m => TypedBuiltinGenT m

A sized built-ins generator that doesn't produce 0 :: Integer, so that one case use div or mod over such integers without the risk of dividing by zero.

genTypedBuiltinAddFailure :: MonadQuote m => TypedBuiltinGenT m

A sized built-ins generator that produces Integers in the (high div 2, high] interval, so that one can use '(+)' over such integers and reliably get an overflow.

genTypedBuiltinMultiplyFailure :: MonadQuote m => TypedBuiltinGenT m

A sized built-ins generator that produces Integers in the (isqrt high, high] interval, so that one can use '(*)' over such integers and reliably get an overflow.

genTypedBuiltinConcatenate :: MonadQuote m => TypedBuiltinGenT m

A sized built-ins generator that produces ByteStrings of such lengths that one can use <> over them without the risk of getting an overflow.

genTypedBuiltinConcatenateFailure :: MonadQuote m => TypedBuiltinGenT m

A sized built-ins generator that produces ByteStrings of such lengths that one can use <> over them and reliably gen an overflow.

type PlcGenT m = GenT (ReaderT (BuiltinGensT m) m)

The type used in generators defined in this module. It's parameterized by a BuiltinGensT which makes it possible to supply different generators of built-in types. For example, genTypedBuiltinDiv never generates a zero, so this generator can be used in order to avoid the divide-by-zero exception. Supplied generators are of arbitrary complexity and can call the currently running generator recursively, for example.

data IterAppValue head arg r

One iterated application of a head to args represented in three distinct ways.

Constructors

IterAppValue

Fields

_iterTerm :: Term TyName Name ()

As a PLC Term.

_iterApp :: IterApp head arg

As an IterApp.

_iterTbv :: TypedBuiltinValue Size r

As a Haskell value.

runPlcT :: Monad m => GenT m Size -> TypedBuiltinGenT m -> PlcGenT m a -> GenT m a

Run a PlcGenT computation by supplying built-ins generators.

genSizeIn :: Monad m => Size -> Size -> GenT m Size

Generate a size from bounds.

genSizeDef :: Monad m => GenT m Size

Generate a size using the default range of [2..4].

genSizeFrom :: Monad m => TypedBuiltin Size a -> GenT m Size

Either return a size taken from a TypedBuiltinSized or generate one using genSizeDef.

genBuiltinSized :: Monad m => GenT m BuiltinSized

Generate a BuiltinSized.

genBuiltin :: Monad m => GenT m Size -> GenT m (BuiltinType size)

Generate a Builtin.

withTypedBuiltinGen :: Monad m => GenT m Size -> (forall a. TypedBuiltin size a -> GenT m c) -> GenT m c

Generate a Builtin and supply its typed version to a continuation.

withCheckedTermGen :: TypedBuiltinGenT Quote -> (forall a. TypedBuiltin Size a -> Maybe (TermOf (Value TyName Name ())) -> GenT Quote c) -> GenT Quote c

Generate a Term along with the value it computes to, having a generator of terms of built-in types.

genIterAppValue :: forall head r m. Monad m => Denotation head Size r -> PlcGenT m (IterAppValue head (Term TyName Name ()) r)

Generate an IterAppValue from a Denotation. If the Denotation has a functional type, then all arguments are generated and supplied to the denotation, the resulting value is forced and if there are any exceptions, then all generated arguments are discarded and another attempt is performed (this process does not loop). Since IterAppValue consists of three components, we 1. grow the Term component by applying it to arguments using Apply 2. grow the IterApp component by appending arguments to its spine 3. feed arguments to the Haskell function

genTerm

Arguments

:: TypedBuiltinGenT Quote
Ground generators of built-ins. The base case of the recursion.

-> DenotationContext
A context to generate terms in. See for example typedBuiltinNames. Gets extended by a variable when an applied lambda is generated.

-> Int
Depth of recursion.

-> TypedBuiltinGenT Quote

Generate a PLC Term of the specified type and the corresponding Haskell value. Generates first-order functions and constants including constant applications. Arguments to functions and BuiltinNames are generated recursively.

genTermLoose :: TypedBuiltinGenT Quote

Generates a Term with rather small values to make out-of-bounds failures less likely. There are still like a half of terms that fail with out-of-bounds errors being evaluated.

withAnyTermLoose :: (forall a. TermOf (TypedBuiltinValue Size a) -> GenT Quote c) -> GenT Quote c

Generate a TypedBuiltin and a TermOf of the corresponding type, attach the TypedBuiltin to the value part of the TermOf and pass that to a continuation.

data Denotation object size r

Haskell denotation of a PLC object. An object can be a BuiltinName or a variable for example.

Constructors

Denotation

Fields

_denotationObject :: object

A PLC object.

_denotationToTerm :: object -> Term TyName Name ()

How to embed the object into a term.

_denotationItself :: a

The denotation of the object. E.g. the denotation of AddInteger is '(+)'.

_denotationScheme :: TypeScheme size a r

The TypeScheme of the object. See sizeIntIntInt for example.

data DenotationContextMember r

A member of a DenotationContext. object is existentially quantified, so the only thing that can be done with it, is turning it into a Term using _denotationToTerm.

Constructors

DenotationContextMember (Denotation object Size r)

newtype DenotationContext

A context of DenotationContextMembers. Each row is a mapping from a type to a list of things that can return that type. For example it can contain a mapping from integer to 1. a bound variable of type integer 2. a bound variable of functional type with the result being integer 3. the AddInteger BuiltinName or any other BuiltinName which returns an integer.

Constructors

DenotationContext

Fields

unDenotationContext :: DMap (TypedBuiltin ()) (Compose [] DenotationContextMember)

denoteVariable :: Name () -> TypedBuiltin size r -> r -> Denotation (Name ()) size r

Get the Denotation of a variable.

denoteTypedBuiltinName :: TypedBuiltinName a r -> a -> Denotation BuiltinName size r

Get the Denotation of a TypedBuiltinName.

insertVariable :: Name () -> TypedBuiltin Size a -> a -> DenotationContext -> DenotationContext

Insert a variable into a DenotationContext.

insertTypedBuiltinName :: TypedBuiltinName a r -> a -> DenotationContext -> DenotationContext

Insert a TypedBuiltinName into a DenotationContext.

typedBuiltinNames :: DenotationContext

A DenotationContext that consists of TypedBuiltinNames.