Language.PlutusCore.Generators.Interesting

Description

Sample generators used for tests.

Documentation

type TermGen size a = GenT Quote (TermOf (TypedBuiltinValue size a))

The type of terms-and-their-values generators.

genOverapplication :: TermGen size Integer

Generates application of a built-in that returns a boolean, immediately saturated afterwards.

lessThanInteger {integer s1} $i1 $i2 {integer s2} $j1 $j2 == if i1 < i2 then j1 else j2

getBuiltinFactorial :: Quote (Term TyName Name ())

i -> product [1 :: Integer .. i] as a PLC term.

/\(s :: size) -> \(i : integer s) ->

let ss = sizeOfInteger {s} i in product {s} ss (enumFromTo {s} (resizeInteger {1} {s} ss 1!1) i)

applyFactorial :: Term TyName Name () -> Size -> Integer -> Term TyName Name ()

Apply some factorial function to its Size and Integer arguments. This function exist, because we have another implementation via dynamic built-ins and want to compare it to the direct implementation from the above.

genFactorial :: TermGen size Integer

Generate a term that computes the factorial of an integer and return it along with the factorial of the corresponding Integer computed on the Haskell side.

genNaiveFib :: TermGen size Integer

Generate a term that computes the ith Fibonacci number and return it along with the corresponding Integer computed on the Haskell side.

genNatRoundtrip :: GenT Quote (TermOf (TypedBuiltinValue size Integer))

Generate an Integer, turn it into a Scott-encoded PLC Nat (see getBuiltinNat), turn that Nat into the corresponding PLC integer using a fold (see getBuiltinFoldNat) defined in terms of generic fix (see getBuiltinFix) and return the result along with the original Integer

genListSum :: TermGen size Integer

Generate a list of Integers, turn it into a Scott-encoded PLC List (see getBuiltinList), sum elements of the list (see getBuiltinSum) and return it along with the sum of the original list.

genIfIntegers :: TermGen size Integer

Generate a boolean and two integers and check whether if b then i1 else i2 means the same thing in Haskell and PLC. Terms are generated using genTermLoose.

fromInterestingTermGens :: (forall a. String -> TermGen size a -> c) -> [c]

Apply a function to all interesting generators and collect the results.