Language.PlutusCore.Generators.Test

Description

This module defines functions useful for testing.

Documentation

data TypeEvalCheckError

The type of errors that can occur during type-eval checking.

Constructors

TypeEvalCheckErrorIllFormed (Error ())

TypeEvalCheckErrorIllEvaled (Value TyName Name ()) (Value TyName Name ())
The former is an expected result of evaluation, the latter -- is an actual one.

data TypeEvalCheckResult

Type-eval checking of a term results in a value of this type.

Constructors

TypeEvalCheckResult

Fields

_termCheckResultType :: NormalizedType TyNameWithKind ()

The type of the term.

_termCheckResultValue :: EvaluationResult

The result of evaluation of the term.

type TypeEvalCheckM = ExceptT TypeEvalCheckError Quote

The monad type-eval checking runs in.

typeEvalCheckBy

Arguments

:: (Term TyName Name () -> EvaluationResult)
An evaluator.

-> TermOf (TypedBuiltinValue Size a)

-> TypeEvalCheckM (TermOf TypeEvalCheckResult)

Type check and evaluate a term and check that the expected result is equal to the actual one.

unsafeTypeEvalCheck :: forall a. TermOf (TypedBuiltinValue Size a) -> Quote (Maybe (TermOf (Value TyName Name ())))

Type check and evaluate a term and check that the expected result is equal to the actual one. Throw an error in case something goes wrong.

sampleTermValue :: TermGen Size a -> IO (TermOf (Value TyName Name ()))

Generate a term using a given generator and check that it's well-typed and evaluates correctly.

sampleProgramValueGolden

Arguments

:: String
folder

-> String
name

-> TermGen Size a
A term generator.

-> IO ()

Generate a pair of files: folder.name.plc and folder.name.plc.golden. The first file contains a term generated by a term generator (wrapped in Program), the second file contains the result of evaluation of the term.

propEvaluate

Arguments

:: (Term TyName Name () -> EvaluationResult)
An evaluator.

-> GenT Quote (TermOf (TypedBuiltinValue Size a))
A term/value generator.

-> Property

A property-based testing procedure for evaluators. Checks whether a term generated along with the value it's supposed to compute to indeed computes to that value according to the provided evaluate.