Language.PlutusCore.MkPlc

Documentation

data VarDecl tyname name a

A "variable declaration", i.e. a name and a type for a variable.

Constructors

VarDecl

Fields

varDeclAnn :: a

varDeclName :: name a

varDeclType :: Type tyname a

data TyVarDecl tyname a

A "type variable declaration", i.e. a name and a kind for a type variable.

Constructors

TyVarDecl

Fields

tyVarDeclAnn :: a

tyVarDeclName :: tyname a

tyVarDeclKind :: Kind a

data TyDecl tyname a

A "type declaration", i.e. a kind for a type.

Constructors

TyDecl

Fields

tyDeclAnn :: a

tyDeclType :: Type tyname a

tyDeclKind :: Kind a

mkVar :: a -> VarDecl tyname name a -> Term tyname name a

Make a Var referencing the given VarDecl.

mkTyVar :: a -> TyVarDecl tyname a -> Type tyname a

Make a TyVar referencing the given TyVarDecl.

tyDeclVar :: TyVarDecl tyname a -> TyDecl tyname a

data Def var val

A definition. Pretty much just a pair with more descriptive names.

Constructors

Def

Fields

defVar :: var

defVal :: val

type TermDef tyname name a = Def (VarDecl tyname name a) (Term tyname name a)

A term definition as a variable.

type TypeDef tyname a = Def (TyVarDecl tyname a) (Type tyname a)

A type definition as a type variable.

data FunctionType tyname ann

The type of a PLC function.

Constructors

FunctionType

Fields

_functionTypeAnn :: ann

An annotation.

_functionTypeDom :: Type tyname ann

The domain of a function.

_functionTypeCod :: Type tyname ann

The codomain of the function.

data FunctionDef tyname name ann

A PLC function.

Constructors

FunctionDef

Fields

_functionDefAnn :: ann

An annotation.

_functionDefName :: name ann

The name of a function.

_functionDefType :: FunctionType tyname ann

The type of the function.

_functionDefTerm :: Term tyname name ann

The definition of the function.

functionTypeToType :: FunctionType tyname ann -> Type tyname ann

Convert a FunctionType to the corresponding Type.

functionDefToType :: FunctionDef tyname name ann -> Type tyname ann

Get the type of a FunctionDef.

functionDefVarDecl :: FunctionDef tyname name ann -> VarDecl tyname name ann

Convert a FunctionDef to a VarDecl. I.e. ignore the actual term.

mkFunctionDef :: ann -> name ann -> Type tyname ann -> Term tyname name ann -> Maybe (FunctionDef tyname name ann)

Make a FunctioDef. Return Nothing if the provided type is not functional.

mkTermLet

Arguments

:: a

-> TermDef tyname name a

-> Term tyname name a
The body of the let, possibly referencing the name.

-> Term tyname name a

Make a "let-binding" for a term.

mkTypeLet

Arguments

:: a

-> TypeDef tyname a

-> Term tyname name a
The body of the let, possibly referencing the name.

-> Term tyname name a

Make a "let-binding" for a type. Note: the body must be a value.

mkIterTyForall :: [TyVarDecl tyname a] -> Type tyname a -> Type tyname a

Universally quantify a list of names.

mkIterTyLam :: [TyVarDecl tyname a] -> Type tyname a -> Type tyname a

Lambda abstract a list of names.

mkIterApp

Arguments

:: a

-> Term tyname name a
f

-> [Term tyname name a]
[ x0 ... xn ]

-> Term tyname name a
[f x0 ... xn ]

Make an iterated application.

mkIterTyFun :: a -> [Type tyname a] -> Type tyname a -> Type tyname a

Make an iterated function type.

mkIterLamAbs :: [VarDecl tyname name a] -> Term tyname name a -> Term tyname name a

Lambda abstract a list of names.

mkIterInst

Arguments

:: a

-> Term tyname name a
a

-> [Type tyname a]
 [ x0 ... xn ]

-> Term tyname name a
{ a x0 ... xn }

Make an iterated instantiation.

mkIterTyAbs :: [TyVarDecl tyname a] -> Term tyname name a -> Term tyname name a

Type abstract a list of names.

mkIterTyApp

Arguments

:: a

-> Type tyname a
f

-> [Type tyname a]
[ x0 ... xn ]

-> Type tyname a
[ f x0 ... xn ]

Make an iterated type application.

mkIterKindArrow :: a -> [Kind a] -> Kind a -> Kind a

Make an iterated function kind.