Language.PlutusCore.Name

Types

type IdentifierState = (IntMap ByteString, Map ByteString Unique, Unique)

An IdentifierState includes a map indexed by Ints as well as a map indexed by ByteStrings. It is used during parsing and renaming.

newtype Unique

A unique identifier

Constructors

Unique

Fields

unUnique :: Int

newtype TypeUnique

The unique of a type-level name.

Constructors

TypeUnique

Fields

unTypeUnique :: Unique

newtype TermUnique

The unique of a term-level name.

Constructors

TermUnique

Fields

unTermUnique :: Unique

class Coercible Unique unique => HasUnique a unique | a -> unique where

Types which have a Unique attached to them, mostly names.

Minimal complete definition

unique

Methods

unique :: Lens' a unique

data Name a

A Name represents variables/names in Plutus Core.

Constructors

Name

Fields

nameAttribute :: a

nameString :: Text

The identifier name, for use in error messages.

nameUnique :: Unique

A Unique assigned to the name during lexing, allowing for cheap comparisons in the compiler.

newtype TyName a

We use a newtype to enforce separation between names used for types and those used for terms.

Constructors

TyName

Fields

unTyName :: Name a

Functions

newIdentifier :: MonadState IdentifierState m => ByteString -> m Unique

This is a naïve implementation of interned identifiers. In particular, it indexes things twice (once by Int, once by ByteString) to ensure fast lookups while lexing and otherwise.

emptyIdentifierState :: IdentifierState

identifierStateFrom :: Unique -> IdentifierState

mapNameString :: (Text -> Text) -> Name a -> Name a

Apply a function to the string representation of a Name.

mapTyNameString :: (Text -> Text) -> TyName a -> TyName a

Apply a function to the string representation of a TyName.

newtypeUnique :: (Wrapped new, HasUnique (Unwrapped new) unique', Coercible unique' unique) => Lens' new unique

The default implementation of HasUnique for newtypes.

data PrettyConfigName

A config that determines how to pretty-print a PLC name.

Constructors

PrettyConfigName

Fields

_pcnShowsUnique :: Bool

Whether to show the Unique of a Name or not.

_pcnShowsAttached :: Bool

Whether to show the "With" part of a name or not. E.g. for TyNameWithKind this flag controls whether the Kind is shown or not.

class HasPrettyConfigName config where

A class of configs from which a PrettyConfigName can be extracted.

Minimal complete definition

toPrettyConfigName

Methods

toPrettyConfigName :: config -> PrettyConfigName

defPrettyConfigName :: PrettyConfigName

The PrettyConfigName used by default: print neither Uniques, nor name attachments.

debugPrettyConfigName :: PrettyConfigName

The PrettyConfigName used for debugging: print Uniques, but not name attachments.