Language.PlutusCore.Normalize

Description

Normalization of PLC entities.

Documentation

data NormalizeTypeT m tyname ann a

The monad transformer that type normalization runs in.

runNormalizeTypeM :: m () -> NormalizeTypeT m tyname ann a -> m a

Run a NormalizeTypeM computation.

runNormalizeTypeDownM :: MonadQuote m => NormalizeTypeT m tyname ann a -> m a

Run a NormalizeTypeM computation without dealing with gas.

runNormalizeTypeGasM :: MonadQuote m => Gas -> NormalizeTypeT (StateT Gas (MaybeT m)) tyname ann a -> m (Maybe a)

Run a gas-consuming NormalizeTypeM computation. Count a single substitution step by subtracting 1 from available gas or fail when there is no available gas.

withExtendedTypeVarEnv :: (HasUnique (tyname ann) TypeUnique, Monad m) => tyname ann -> NormalizedType tyname ann -> NormalizeTypeT m tyname ann a -> NormalizeTypeT m tyname ann a

Locally extend a TypeVarEnv in a NormalizeTypeM computation.

normalizeTypeM :: (HasUnique (tyname ann) TypeUnique, MonadQuote m) => Type tyname ann -> NormalizeTypeT m tyname ann (NormalizedType tyname ann)

Normalize a Type in the NormalizeTypeM monad.

normalizeTypeDown :: (HasUnique (tyname ann) TypeUnique, MonadQuote m) => Type tyname ann -> m (NormalizedType tyname ann)

substituteNormalizeTypeM

Arguments

:: (HasUnique (tyname ann) TypeUnique, MonadQuote m)

=> NormalizedType tyname ann
ty

-> tyname ann
name

-> Type tyname ann
body

-> NormalizeTypeT m tyname ann (NormalizedType tyname ann)
NORM ([ty / name] body)

Substitute a type for a variable in a type and normalize in the NormalizeTypeM monad.

normalizeTypesIn :: (HasUnique (tyname ann) TypeUnique, MonadQuote m) => Term tyname name ann -> NormalizeTypeT m tyname ann (Term tyname name ann)

Normalize every Type in a Term.