Language.PlutusCore.Pretty

Basic types and functions

data Doc ann

class Pretty a where

Minimal complete definition

pretty

Methods

pretty :: a -> Doc ann

prettyList :: [a] -> Doc ann

class PrettyBy config a where

Overloaded configurable conversion to Doc. I.e. like Pretty, but parameterized by a config. This class is interoperable with the Pretty class via PrettyConfigIgnore and PrettyConfigAttatch.

Methods

prettyBy :: config -> a -> Doc ann

prettyBy :: DefaultPrettyBy config a => config -> a -> Doc ann

newtype PrettyConfigIgnore a

A newtype wrapper around a which point is to provide a 'PrettyBy config' instance for anything that has a Pretty instance.

Constructors

PrettyConfigIgnore

Fields

unPrettyConfigIgnore :: a

data PrettyConfigAttach config a

A config together with some value. The point is to provide a Pretty instance for anything that has a 'PrettyBy config' instance.

Constructors

PrettyConfigAttach config a

docString :: Doc a -> String

Render a Doc as String.

docText :: Doc a -> Text

Render a Doc as Text.

prettyString :: Pretty a => a -> String

Render a value as String.

prettyText :: Pretty a => a -> Text

Render a value as strict Text.

prettyStringBy :: PrettyBy config a => config -> a -> String

Render a value as String.

prettyTextBy :: PrettyBy config a => config -> a -> Text

Render a value as strict Text.

Defaults

prettyPlcDef :: PrettyPlc a => a -> Doc ann

Pretty-print a value in the default mode using the classic view.

prettyPlcDefString :: PrettyPlc a => a -> String

Render a value to String in the default mode using the classic view.

prettyPlcDefText :: PrettyPlc a => a -> Text

Render a value to Text in the default mode using the classic view.

Global configuration

data CondensedErrors

Whether to pretty-print PLC errors in full or with some information omitted.

Constructors

CondensedErrorsYes

CondensedErrorsNo

newtype PrettyConfigPlcOptions

Options for pretty-printing PLC entities.

Constructors

PrettyConfigPlcOptions

Fields

_pcpoCondensedErrors :: CondensedErrors

data PrettyConfigPlcStrategy

Strategy for pretty-printing PLC entities.

Constructors

PrettyConfigPlcClassic (PrettyConfigClassic PrettyConfigName)

PrettyConfigPlcReadable (PrettyConfigReadable PrettyConfigName)

data PrettyConfigPlc

Global configuration used for pretty-printing PLC entities.

Constructors

PrettyConfigPlc

Fields

_pcpOptions :: PrettyConfigPlcOptions

_pcpStrategy :: PrettyConfigPlcStrategy

type PrettyPlc = PrettyBy PrettyConfigPlc

The "pretty-printable PLC entity" constraint.

defPrettyConfigPlcOptions :: PrettyConfigPlcOptions

The PrettyConfigPlcOptions used by default: print errors in full.

defPrettyConfigPlcClassic :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used by default: use the classic view and print neither Uniques, nor name attachments.

debugPrettyConfigPlcClassic :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used for debugging: use the classic view and print Uniques, but not name attachments.

defPrettyConfigPlcReadable :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used by default and for readability: use the refined view and print neither Uniques, nor name attachments.

debugPrettyConfigPlcReadable :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used for debugging and readability: use the refined view and print Uniques, but not name attachments.

Custom functions for PLC types.

prettyPlcClassicDef :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value in the default mode using the classic view.

prettyPlcClassicDebug :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value in the debug mode using the classic view.

prettyPlcReadableDef :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value in the default mode using the readable view.

prettyPlcReadableDebug :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value in the debug mode using the readable view.

prettyPlcCondensedErrorBy :: PrettyPlc a => (PrettyConfigPlcOptions -> PrettyConfigPlc) -> a -> Doc ann

Pretty-print a PLC value using the condensed way (see CondensedErrors) of pretty-printing PLC errors (in case there are any).

prettyPlcCondensedErrorClassicString :: PrettyPlc a => a -> String

Render an error to String in the condensed manner using the classic view.

Names

data PrettyConfigName

A config that determines how to pretty-print a PLC name.

Constructors

PrettyConfigName

Fields

_pcnShowsUnique :: Bool

Whether to show the Unique of a Name or not.

_pcnShowsAttached :: Bool

Whether to show the "With" part of a name or not. E.g. for TyNameWithKind this flag controls whether the Kind is shown or not.

class HasPrettyConfigName config where

A class of configs from which a PrettyConfigName can be extracted.

Minimal complete definition

toPrettyConfigName

Methods

toPrettyConfigName :: config -> PrettyConfigName

defPrettyConfigName :: PrettyConfigName

The PrettyConfigName used by default: print neither Uniques, nor name attachments.

debugPrettyConfigName :: PrettyConfigName

The PrettyConfigName used for debugging: print Uniques, but not name attachments.

Classic view

newtype PrettyConfigClassic configName

Configuration for the classic pretty-printing.

Constructors

PrettyConfigClassic

Fields

_pccConfigName :: configName

type PrettyClassicBy configName = PrettyBy (PrettyConfigClassic configName)

The "classically pretty-printable" constraint.

type PrettyClassic = PrettyClassicBy PrettyConfigName

prettyClassicDef :: PrettyClassic a => a -> Doc ann

Pretty-print a value in the default mode using the classic view.

Readable view

data RenderContext

A context an expression is rendering in.

Constructors

RenderContext

Fields

_rcFixity :: Fixity

_rcDirection :: Direction

data ShowKinds

Constructors

ShowKindsYes

ShowKindsNo

data PrettyConfigReadable configName

Configuration for the readable pretty-printing.

Constructors

PrettyConfigReadable

Fields

_pcrConfigName :: configName

_pcrRenderContext :: RenderContext

_pcrShowKinds :: ShowKinds

type PrettyReadableBy configName = PrettyBy (PrettyConfigReadable configName)

The "readably pretty-printable" constraint.

type PrettyReadable = PrettyReadableBy PrettyConfigName

topPrettyConfigReadable :: configName -> ShowKinds -> PrettyConfigReadable configName

A PrettyConfigReadable with the fixity specified to topApp.

botPrettyConfigReadable :: configName -> ShowKinds -> PrettyConfigReadable configName

A PrettyConfigReadable with the fixity specified to botApp.

Orphan instances

Pretty (Kind a)

Instance details

Methods

pretty :: Kind a -> Doc ann

prettyList :: [Kind a] -> Doc ann

Pretty (Constant a)

Instance details

Methods

pretty :: Constant a -> Doc ann

prettyList :: [Constant a] -> Doc ann

Pretty (Builtin a)

Instance details

Methods

pretty :: Builtin a -> Doc ann

prettyList :: [Builtin a] -> Doc ann

PrettyClassic (Type tyname a) => Pretty (Type tyname a)

Instance details

Methods

pretty :: Type tyname a -> Doc ann

prettyList :: [Type tyname a] -> Doc ann

PrettyClassic (Program tyname name a) => Pretty (Program tyname name a)

Instance details

Methods

pretty :: Program tyname name a -> Doc ann

prettyList :: [Program tyname name a] -> Doc ann

PrettyClassic (Term tyname name a) => Pretty (Term tyname name a)

Instance details

Methods

pretty :: Term tyname name a -> Doc ann

prettyList :: [Term tyname name a] -> Doc ann