Language.PlutusCore.Quote

Documentation

runQuoteT :: Monad m => QuoteT m a -> m a

Run a quote from an empty identifier state. Note that the resulting term cannot necessarily be safely combined with other terms - that should happen inside QuoteT.

runQuote :: Quote a -> a

See runQuoteT.

freshUnique :: MonadQuote m => m Unique

Get a fresh Unique.

freshName :: Monad m => a -> Text -> QuoteT m (Name a)

Get a fresh Name, given the annotation and the Text name.

freshTyName :: Monad m => a -> Text -> QuoteT m (TyName a)

Get a fresh TyName, given the annotation and the Text name.

freshenName :: Monad m => Name a -> QuoteT m (Name a)

Make a copy of the given Name that is distinct from the old one.

freshenTyName :: Monad m => TyName a -> QuoteT m (TyName a)

Make a copy of the given TyName that is distinct from the old one.

markNonFresh :: MonadQuote m => Unique -> m ()

Mark a given Unique (and implicitly all Uniques less than it) as used, so they will not be generated in future.

markNonFreshTerm :: (HasUnique (tyname a) TypeUnique, HasUnique (name a) TermUnique, MonadQuote m) => Term tyname name a -> m ()

Marks all the Uniques in a term as used, so they will not be generated in future. Useful if you have a term which was not generated in Quote.

markNonFreshType :: (HasUnique (tyname a) TypeUnique, MonadQuote m) => Type tyname a -> m ()

Marks all the Uniques in a type as used, so they will not be generated in future. Useful if you have a type which was not generated in Quote.

markNonFreshProgram :: (HasUnique (tyname a) TypeUnique, HasUnique (name a) TermUnique, MonadQuote m) => Program tyname name a -> m ()

Marks all the Uniques in a program as used, so they will not be generated in future. Useful if you have a program which was not generated in Quote.

data QuoteT m a

The "quotation" monad transformer. Within this monad you can do safe construction of PLC terms using quasiquotation, fresh-name generation, and parsing.

type Quote = QuoteT Identity

A non-transformer version of QuoteT.

class Monad m => MonadQuote m

A monad that allows lifting of quoted expressions.

type FreshState = Unique

The state contains the "next" Unique that should be used for a name

liftQuote :: MonadQuote m => Quote a -> m a