Language.PlutusCore.StdLib.Meta

Description

Functions that generate Plutus Core terms from Haskell values and vice versa.

Documentation

getBuiltinIntegerToNat :: Integer -> Quote (Term TyName Name ())

Convert an Integer to a nat. TODO: convert PLC's integer to nat instead.

getEitherToBuiltinSum :: Type TyName () -> Type TyName () -> Either (Term TyName Name ()) (Term TyName Name ()) -> Quote (Term TyName Name ())

Convert a Haskell Either to a PLC sum.

getListToBuiltinList :: Type TyName () -> [Term TyName Name ()] -> Quote (Term TyName Name ())

Convert a Haskell list of Terms to a PLC list.