Language.PlutusCore.StdLib.Meta.Data.Tuple

Description

tuples of various sizees and related functions.

Documentation

data Tuple ann

A Plutus Core tuple.

Constructors

Tuple

Fields

_tupleElementTypes :: [Type TyName ann]

The types of elements of a tuple.

_tupleTerm :: Term TyName Name ann

A term representation of the tuple.

getTupleType :: MonadQuote m => ann -> Tuple ann -> m (Type TyName ann)

Get the type of a Tuple.

getTupleType _ (Tuple [a1, ... , an] _) = all r. (a1 -> ... -> an -> r) -> r

tupleTypeTermAt :: MonadQuote m => ann -> Int -> Tuple ann -> m (Type TyName ann, Term TyName Name ann)

Get the type of the ith element of a Tuple along with the element itself.

tupleTypeTermAt _ i (Tuple [a0, ... , an] term) =
    (ai, term {ai} (\(x0 : a0) ... (xn : an) -> xi))

tupleTermAt :: MonadQuote m => ann -> Int -> Tuple ann -> m (Term TyName Name ann)

Get the ith element of a Tuple.

tupleDefAt :: MonadQuote m => ann -> Int -> Name ann -> Tuple ann -> m (TermDef TyName Name ann)

Get the ith element of a Tuple as a TermDef.

bindTuple :: MonadQuote m => ann -> [Name ann] -> Tuple ann -> Term TyName Name ann -> m (Term TyName Name ann)

Bind all elements of a Tuple inside a Term.

bindTuple _ [x_1, ... , x_n] (Tuple [a1, ... , an] term) body =
    (\(tup : all r. (a_1 -> ... -> a_n -> r) -> r) ->
      let x_1 = _1 tup
          ...
          x_n = _n tup
        in body
    ) term

getBuiltinProdN :: MonadQuote m => Int -> m (Type TyName ())

Given an arity n, create the n-ary product type.

(T_1 :: *) .. (T_n :: *) . all (R :: *) . (T_1 -> .. -> T_n -> R) -> R

getBuiltinProdNConstructor :: MonadQuote m => Int -> m (Term TyName Name ())

Given an arity n, create the constructor for n-ary products.

    /(T_1 :: *) .. (T_n :: *) .
        (arg_1 : T_1) .. (arg_n : T_n) .
            /(R :: *).
                (case : T_1 -> .. -> T_n -> R) -> case arg_1 .. arg_n

getBuiltinProdNAccessor :: MonadQuote m => Int -> Int -> m (Term TyName Name ())

Given an arity n and an index i, create a function for accessing the i'th component of a n-tuple.

    /(T_1 :: *) .. (T_n :: *) .
        (tuple : all (R :: *) . (T_1 -> .. -> T_n -> R) -> R)) .
            tuple {T_i} ((arg_1 : T_1) .. (arg_n : T_n) . arg_i)

getSpineToTuple :: MonadQuote m => ann -> [(Type TyName ann, Term TyName Name ann)] -> m (Tuple ann)

Convert a Haskell spine of Terms to a PLC Tuple.

getSpineToTuple _ [(a1, x1), ... , (an, xn)] =
    Tuple [a1, ... , an] (/\(r :: *) -> \(f :: a1 -> ... -> an -> r) -> f x1 ... xn)