Language.PlutusCore.Subst

Documentation

substTerm :: (tyname a -> Maybe (Type tyname a)) -> (name a -> Maybe (Term tyname name a)) -> Term tyname name a -> Term tyname name a

Naively substitute names using the given functions (i.e. do not account for scoping).

substTy :: (tyname a -> Maybe (Type tyname a)) -> Type tyname a -> Type tyname a

Naively substitute names using the given function (i.e. do not account for scoping).

fvTerm :: Ord (name a) => Term tyname name a -> Set (name a)

Get all the free term variables in a term.

ftvTerm :: Ord (tyname a) => Term tyname name a -> Set (tyname a)

Get all the free type variables in a term.

ftvTy :: Ord (tyname a) => Type tyname a -> Set (tyname a)

Get all the free type variables in a type.