Language.PlutusCore.View

Description

Various views of PLC entities.

Documentation

data IterApp head arg

A function (called "head") applied to a list of arguments (called "spine").

Constructors

IterApp

Fields

_iterAppHead :: head

_iterAppSpine :: [arg]

type TermIterApp tyname name a = IterApp (Term tyname name a) (Term tyname name a)

An iterated application of a Term to a list of Terms.

type PrimIterApp tyname name a = IterApp StagedBuiltinName (Value tyname name a)

An iterated application of a BuiltinName to a list of Values.

constantAsInteger :: Constant a -> Maybe Integer

View a Constant as an Integer.

constantAsStagedBuiltinName :: Builtin a -> StagedBuiltinName

View a Constant as a StagedBuiltinName.

termAsBuiltin :: Term tyname name a -> Maybe (Builtin a)

View a Term as a Constant.

termAsTermIterApp :: Term tyname name a -> TermIterApp tyname name a

An iterated application of a Term to a list of Terms.

termAsPrimIterApp :: Term tyname name a -> Maybe (PrimIterApp tyname name a)

View a Term as an iterated application of a BuiltinName to a list of Values.

isTermValue :: Term tyname name a -> Bool

Check whether a Term is a Value. The term is assumed to be valid.